Move and reword auto mods setting
